AI project risk chain/llm/llm/Supply-Chain-Risk-Analyzer/k/supply_chain_risk_analyzer/requirements.txt
google-genai>=1.0.0
tavily-python>=0.3.0
groq>=0.9.0
streamlit>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
plotly>=5.18.0
pandas>=2.0.0
